refactor: Interfaces e funções espalhadas pelo projeto foram movidas para a pasta shared
