+++ web/src/app/shared/components/button/button.component.ts
@@ -1,15 +1,15 @@
 import { Component, input } from '@angular/core';
-import { Button } from "../table/table.component";
 import { ButtonModule } from "primeng/button";
+import { Button } from "./models/button";
 
 @Component({
-  selector: 'app-button',
-  standalone: true,
+	selector: 'app-button',
+	standalone: true,
 	imports: [
 		ButtonModule
 	],
-  templateUrl: './button.component.html',
-  styleUrl: './button.component.scss'
+	templateUrl: './button.component.html',
+	styleUrl: './button.component.scss'
 })
 export class ButtonComponent {
 	action = input.required<Button>()
