+++ web/src/app/shared/components/button/models/button.ts
@@ -1,0 +1,11 @@
+import { ColorSeverity } from "../../../models/color-severity";
+
+export interface Button {
+	icon?: string;
+	label?: string;
+	condition?: boolean;
+	size?: 'small' | 'large';
+	severity?: ColorSeverity;
+	theme?: 'raised' | 'outlined' | 'plain';
+	click: () => void;
+}
