+++ web/src/app/shared/functions/get-error.message.ts
@@ -1,0 +1,12 @@
+import { FormControl } from "@angular/forms";
+
+export const getErrorMessage = (control: FormControl) => {
+	const customError = control?.errors?.["customError"];
+
+	if (control?.errors?.["required"])
+		return "Campo é obrigatório";
+	else if (customError)
+		return customError;
+
+	return "";
+}
