+++ web/src/app/shared/models/base-select-option.ts
@@ -1,0 +1,4 @@
+export interface BaseSelectOption {
+	id: number;
+	name: string;
+}
