+++ web/src/app/shared/models/dialog-component.info.ts
@@ -1,0 +1,3 @@
+import { Generic } from "./generic";
+
+export type DialogComponentInfo<T> = T extends { dialogInfo: Generic } ? T["dialogInfo"] : any;
