+++ web/src/app/shared/models/generic.ts
@@ -1,0 +1,1 @@
+export type Generic = Record<string, any>;
