+++ web/src/app/shared/models/identifiable.ts
@@ -1,0 +1,3 @@
+export interface Identifiable {
+	id: number;
+}
