+++ web/src/app/shared/models/table-columns-fn.ts
@@ -1,0 +1,3 @@
+import { TableColumn } from "../components/table/models/table-column";
+
+export type TableColumnsFn<T> = (column: T) => TableColumn[];
